feat: Add new support functions and perform cleanups
